+++ .github/workflows/apisix-conformance-test.yml
@@ -34,6 +34,8 @@ permissions:
 
 jobs:
   conformance-test:
+    env:
+      CONFORMANCE_TEST_REPORT_OUTPUT: /tmp/apisix-ingress-controller-conformance-report.yaml
     timeout-minutes: 60
     strategy:
       matrix:
@@ -61,6 +63,10 @@ jobs:
           registry: ${{ secrets.DOCKER_REGISTRY }}
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Install kind
+        run: |
+          go install sigs.k8s.io/kind@v0.23.0
 
       - name: Build images
         env:
@@ -97,27 +103,31 @@ jobs:
         env:
           PROVIDER_TYPE: ${{ matrix.provider_type }}
         run: |
-          make conformance-test-standalone
+          make conformance-test
       
-      - name: Get Logs from api7-ingress-controller
+      - name: Show Conformance Report
+        shell: bash
+        run: |
+          cat ${CONFORMANCE_TEST_REPORT_OUTPUT}
+
+      - name: Get Logs from apisix-ingress-controller
         shell: bash
         run: |
           export KUBECONFIG=/tmp/apisix-ingress-cluster.kubeconfig
           kubectl logs -n apisix-conformance-test -l app=apisix-ingress-controller
 
       - name: Upload Gateway API Conformance Report
-        if: ${{ github.event_name == 'push' }}
         uses: actions/upload-artifact@v4
         with:
           name: apisix-ingress-controller-conformance-report-${{ matrix.provider_type }}.yaml
-          path: apisix-ingress-controller-conformance-report.yaml
+          path: ${{ env.CONFORMANCE_TEST_REPORT_OUTPUT }}
 
       - name: Format Conformance Test Report
         if: ${{ github.event_name == 'pull_request' }}
         run: |
           echo '# conformance test report - ${{ matrix.provider_type }} mode' > report.md
           echo '```yaml' >> report.md
-          cat apisix-ingress-controller-conformance-report.yaml >> report.md
+          cat ${CONFORMANCE_TEST_REPORT_OUTPUT} >> report.md
           echo '```' >> report.md
 
       - name: Report Conformance Test Result to PR Comment
