+++ test/conformance/api7ee/suite_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package conformance
+package api7ee
 
 import (
 	"context"
@@ -32,7 +32,6 @@ import (
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	"github.com/apache/apisix-ingress-controller/test/e2e/framework"
-	"github.com/apache/apisix-ingress-controller/test/e2e/scaffold"
 )
 
 var gatewayClassName = "apisix"
@@ -124,8 +123,7 @@ func TestMain(m *testing.M) {
 	RegisterFailHandler(Fail)
 	f := framework.NewFramework()
 
-	// init newDeployer function
-	scaffold.NewDeployer = scaffold.NewAPISIXDeployer
+	f.BeforeSuite()
 
 	// Check and delete specific namespaces if they exist
 	kubectl := k8s.NewKubectlOptions("", "", "default")
@@ -138,23 +136,22 @@ func TestMain(m *testing.M) {
 	k8s.CreateNamespace(GinkgoT(), kubectl, namespace)
 	defer k8s.DeleteNamespace(GinkgoT(), kubectl, namespace)
 
-	adminkey := getEnvOrDefault("APISIX_ADMIN_KEY", "edd1c9f034335f136f87ad84b625c8f1")
-	controllerName := "apisix.apache.org/apisix-ingress-controller"
-	s := scaffold.NewScaffold(scaffold.Options{
-		ControllerName:    controllerName,
-		SkipHooks:         true,
-		APISIXAdminAPIKey: adminkey,
-	})
+	gatewayGroupId := f.CreateNewGatewayGroupWithIngress()
+	adminKey := f.GetAdminKey(gatewayGroupId)
 
-	s.Deployer.DeployDataplane(scaffold.DeployDataplaneOptions{
-		AdminKey:          adminkey,
-		Namespace:         namespace,
-		SkipCreateTunnels: true,
-		ServiceType:       "LoadBalancer",
-		ServiceHTTPPort:   80,
-		ServiceHTTPSPort:  443,
+	svc := f.DeployGateway(&framework.API7DeployOptions{
+		Namespace:              namespace,
+		GatewayGroupID:         gatewayGroupId,
+		DPManagerEndpoint:      framework.DPManagerTLSEndpoint,
+		SetEnv:                 true,
+		SSLKey:                 framework.TestKey,
+		SSLCert:                framework.TestCert,
+		TLSEnabled:             true,
+		ForIngressGatewayGroup: true,
+		ServiceType:            "LoadBalancer",
+		ServiceHTTPPort:        80,
+		ServiceHTTPSPort:       443,
 	})
-	svc := s.GetDataplaneService()
 
 	if len(svc.Status.LoadBalancer.Ingress) == 0 {
 		Fail("No LoadBalancer found for the service")
@@ -163,25 +160,24 @@ func TestMain(m *testing.M) {
 	address := svc.Status.LoadBalancer.Ingress[0].IP
 
 	f.DeployIngress(framework.IngressDeployOpts{
-		ControllerName:     controllerName,
-		Namespace:          namespace,
-		StatusAddress:      address,
-		InitSyncDelay:      20 * time.Minute,
-		ProviderType:       framework.ProviderType,
-		ProviderSyncPeriod: 1 * time.Hour,
+		ControllerName: "apisix.apache.org/apisix-ingress-controller",
+		Namespace:      namespace,
+		StatusAddress:  address,
+		InitSyncDelay:  1 * time.Minute,
+		ProviderType:   "api7ee",
 	})
-
-	adminEndpoint := fmt.Sprintf("http://%s.%s:9180", svc.Name, namespace)
 
 	defaultGatewayProxyOpts = GatewayProxyOpts{
 		StatusAddress: address,
-		AdminKey:      adminkey,
-		AdminEndpoint: adminEndpoint,
+		AdminKey:      adminKey,
+		AdminEndpoint: framework.DashboardTLSEndpoint,
 	}
 
 	patchGatewaysForConformanceTest(context.Background(), f.K8sClient)
 
 	code := m.Run()
+
+	f.AfterSuite()
 
 	os.Exit(code)
 }
@@ -261,12 +257,4 @@ func patchGatewaysForConformanceTest(ctx context.Context, k8sClient client.Clien
 			}
 		}
 	}()
-}
-
-// getEnvOrDefault returns environment variable value or default
-func getEnvOrDefault(key, defaultValue string) string {
-	if value := os.Getenv(key); value != "" {
-		return value
-	}
-	return defaultValue
 }
