+++ test/conformance/conformance_test.go
@@ -15,52 +15,34 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//go:build conformance
-// +build conformance
-
 package conformance
 
 import (
-	"flag"
 	"os"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"sigs.k8s.io/gateway-api/conformance"
 	conformancev1 "sigs.k8s.io/gateway-api/conformance/apis/v1"
 	"sigs.k8s.io/gateway-api/conformance/tests"
+	"sigs.k8s.io/gateway-api/conformance/utils/flags"
 	"sigs.k8s.io/gateway-api/conformance/utils/suite"
-	"sigs.k8s.io/gateway-api/pkg/features"
 	"sigs.k8s.io/yaml"
 )
 
+// https://github.com/kubernetes-sigs/gateway-api/blob/5c5fc388829d24e8071071b01e8313ada8f15d9f/conformance/utils/suite/suite.go#L358.  SAN includes '*'
 var skippedTestsForSSL = []string{
-	// Reason: https://github.com/kubernetes-sigs/gateway-api/blob/5c5fc388829d24e8071071b01e8313ada8f15d9f/conformance/utils/suite/suite.go#L358.  SAN includes '*'
 	tests.HTTPRouteHTTPSListener.ShortName,
 	tests.HTTPRouteRedirectPortAndScheme.ShortName,
 }
 
 // TODO: HTTPRoute hostname intersection and listener hostname matching
 
-var gatewaySupportedFeatures = []features.FeatureName{
-	features.SupportGateway,
-	features.SupportHTTPRoute,
-	// features.SupportHTTPRouteMethodMatching,
-	// features.SupportHTTPRouteResponseHeaderModification,
-	// features.SupportHTTPRouteRequestMirror,
-	// features.SupportHTTPRouteBackendRequestHeaderModification,
-	// features.SupportHTTPRouteHostRewrite,
-}
-
 func TestGatewayAPIConformance(t *testing.T) {
-	flag.Parse()
-
 	opts := conformance.DefaultOptions(t)
 	opts.Debug = true
 	opts.CleanupBaseResources = true
 	opts.GatewayClassName = gatewayClassName
-	opts.SupportedFeatures = sets.New(gatewaySupportedFeatures...)
 	opts.SkipTests = skippedTestsForSSL
 	opts.Implementation = conformancev1.Implementation{
 		Organization: "APISIX",
@@ -68,7 +50,6 @@ func TestGatewayAPIConformance(t *testing.T) {
 		URL:          "https://github.com/apache/apisix-ingress-controller.git",
 		Version:      "v2.0.0",
 	}
-	opts.ConformanceProfiles = sets.New(suite.GatewayHTTPConformanceProfileName)
 
 	cSuite, err := suite.NewConformanceTestSuite(opts)
 	require.NoError(t, err)
@@ -80,7 +61,6 @@ func TestGatewayAPIConformance(t *testing.T) {
 		t.Fatalf("failed to run the gateway conformance test suite: %v", err)
 	}
 
-	const reportFileName = "apisix-ingress-controller-conformance-report.yaml"
 	report, err := cSuite.Report()
 	if err != nil {
 		t.Fatalf("failed to get the gateway conformance test report: %v", err)
@@ -90,6 +70,10 @@ func TestGatewayAPIConformance(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to marshal the gateway conformance test report: %v", err)
 	}
-	// Save report in the root of the repository, file name is in .gitignore.
-	require.NoError(t, os.WriteFile("../../"+reportFileName, rawReport, 0o600))
+	f, err := os.Create(*flags.ReportOutput)
+	require.NoError(t, err)
+	defer func() { _ = f.Close() }()
+
+	_, err = f.Write(rawReport)
+	require.NoError(t, err)
 }
